refactor: split Queue and Worker
Moves QueueWorker to a separate file, and prepares Queue for handling messages from workers.

queue_worker.rs:
- move QueueWorker to `queue_worker.rs` .
- add event sender param to QueueWorker.

queue.rs:
- rename: Queue `sender/receiver` channels to `item_sender/item_receiver`.
- add: QueueSize struct for interacting with `Queue.size` from main event listener loop.
- add: `WorkerComplete` variant to QueueMsg enum.
- add: embed values to track file hash in QueueMsg enum.
