+++ src-tauri/src/img_cache/mod.rs
@@ -2,3 +2,4 @@ pub mod cache;
 pub mod hash;
 pub mod img;
 pub mod queue;
+mod queue_worker;
