+++ src-tauri/src/img_cache/queue.rs
@@ -1,18 +1,20 @@
 use log::{debug, error, info};
 use std::{
-    io::ErrorKind, path::PathBuf, sync::{
+    // io::ErrorKind,
+    collections::{HashMap, HashSet}, path::PathBuf, sync::{
         atomic::{AtomicUsize, Ordering},
         Arc,
     }
 };
 use tokio::{
-    fs,
+    // fs,
     sync::mpsc,
-    task,
+    // task,
 };
 
-use crate::{img_cache::{cache, img}, ipc::send};
-use crate::get_db;
+use crate::img_cache::queue_worker::QueueWorker;
+// use crate::{img_cache::{cache, img, queue_worker::{handle_create_thumbnail_result, QueueWorker}}, ipc::send};
+// use crate::get_db;
 
 #[derive(Debug, Clone)]
 pub struct QueueItem {
@@ -23,37 +25,38 @@ pub struct QueueItem {
 
 #[derive(Debug, Clone)]
 pub enum QueueMsg {
-    ImageAdded,
-    WorkerComplete,
+    ImageAdded { file_hash: String },
+    ImageFailed { file_hash: String },
+    WorkerComplete { file_hash: String },
 }
 
 #[derive(Debug)]
 pub struct Queue {
-    sender: mpsc::UnboundedSender<QueueItem>,
+    item_sender: mpsc::UnboundedSender<QueueItem>,
     size: Arc<AtomicUsize>,
     event_sender: mpsc::UnboundedSender<QueueMsg>,
 }
 
 impl Queue {
     pub fn new() -> (Self, mpsc::UnboundedReceiver<QueueItem>, mpsc::UnboundedReceiver<QueueMsg>) {
-        let (sender, receiver) = mpsc::unbounded_channel();
+        let (item_sender, item_receiver) = mpsc::unbounded_channel();
         let (event_sender, event_receiver) = mpsc::unbounded_channel();
         let size = Arc::new(AtomicUsize::new(0));
 
-        let queue = Queue { sender, size, event_sender };
+        let queue = Queue { item_sender, size, event_sender };
 
-        (queue, receiver, event_receiver)
+        (queue, item_receiver, event_receiver)
     }
 
-    // TODO: FUTURE: enqueue allows duplicates to be added to queue,
-    // causing thumbnails to be generated twice.
-    // this should now be mostly resolved by directory locking in frontend.
-    // revisit with REFACTOR_IMAGE_PIPELINE
+    // enqueue an item and increment the queue size
     pub async fn enqueue(&self, item: QueueItem) -> Result<(), String> {
-        match self.sender.send(item) {
+        let event_payload = item_hash(&item);
+        match self.item_sender.send(item) {
             Ok(_) => {
                 self.size.fetch_add(1, Ordering::SeqCst);
-                let _ = self.event_sender.send(QueueMsg::ImageAdded);
+                let _ = self.event_sender.send(
+                    QueueMsg::ImageAdded{ file_hash: event_payload }
+                );
                 Ok(())
             }
             Err(_) => Err("Failed to enqueue item".to_string()),
@@ -65,113 +68,20 @@ impl Queue {
     }
 }
 
-// Worker that processes the queue
-pub struct QueueWorker;
-
-impl QueueWorker {
-    pub async fn start(
-        mut receiver: mpsc::UnboundedReceiver<QueueItem>,
-        size: Arc<AtomicUsize>,
-    ) {
-        // check base cache directory
-        let cache_path = match cache::get_cache_path_async().await {
-            Ok(path) => path,
-            Err(e) => {
-                error!("QueueWorker failed to get or create cache path: {e}");
-                return;
-            }
-        };
-
-        while let Some(item) = receiver.recv().await {
-            debug!("Generate thumbnail: {}", item.full_path.to_string_lossy());
-
-            let mut thumb_path = cache_path.clone();
-            thumb_path.push(&item.path_hash);
-
-            // ensure the specific item's hash directory exists
-            match fs::create_dir_all(&thumb_path).await {
-                Ok(_) => {},
-                Err(e) if e.kind() == ErrorKind::AlreadyExists => {},
-                Err(e) => {
-                    // TODO: granular error handling
-                    error!("Failed to create item-specific directory {thumb_path:?}: {e}");
-                    size.fetch_sub(1, Ordering::SeqCst);
-                    continue;
-                }
-            }
-
-            thumb_path.push(&item.file_hash);
-            thumb_path.set_extension("webp");
-
-            let org_path = item.full_path.clone();
-            let result = task::spawn_blocking(move || {
-                img::create_thumbnail(org_path.as_path(), &thumb_path)
-            })
-            .await; // awaits JoinHandle from spawn_blocking
-
-            match result {
-                Ok(Ok(())) => {
-                    debug!("Thumbnail generated successfully for {}", item.full_path.to_string_lossy());
-
-                    if let Err(e) = get_db().insert_hash(
-                        &item.path_hash,
-                        &item.file_hash,
-                        &item.full_path.to_string_lossy()
-                    ) {
-                        error!("Failed to insert hash for {}, {}", item.full_path.to_string_lossy(), e);
-                    }
-                }
-                Ok(Err(e)) => {
-                    let failed_img_path = item.full_path.to_string_lossy();
-                    match e {
-                        image::ImageError::Unsupported(unsupported_error) => {
-                            error!("image failed to generate thumbnail for file: {failed_img_path}");
-                            error!("image error: {unsupported_error}");
-                            send::info_window_msg(&format!("Failed to create thumbnail for: {failed_img_path}"));
-                            send::info_window_msg(&format!("Reason: {unsupported_error}"));
-                        }
-                        _ => {
-                            error!("Error generating thumbnail for {failed_img_path}: {e}");
-                        }
-                    }
-                }
-                Err(join_error) => {
-                    error!("Spawn blocking task for {} failed: {:?}", item.full_path.to_string_lossy(), join_error);
-
-                    if join_error.is_panic() {
-                        error!("try_into_panic(): true");
-                    }
-                    if join_error.is_cancelled() {
-                        error!("is_cancelled(): true");
-                    }
-                }
-            }
-
-            let remaining = size.fetch_sub(1, Ordering::SeqCst).saturating_sub(1);
-            if remaining % 10 == 0 || remaining < 10 {
-                if remaining > 0 {
-                    send::info_window_msg(&format!("{remaining:?} thumbnails in queue."));
-                }
-                else {
-                    if let Some(parent) = item.full_path.parent() {
-                        // TODO: readability TODO_QUEUE_STATUS_PARAMS
-                        // update to use enum/type params
-                        // sending: redraw=true, queue_empty=true
-                        send::queue_status(parent, &true, &true);
-                    }
-                    send::info_window_msg("Thumbnails generated.");
-                }
-                info!("{remaining:?} images remaining in queue.");
-            }
-        }
-        info!("Worker shutting down");
-    }
+struct QueueSize {
+    size: Arc<AtomicUsize>
+}
+impl QueueSize {
+    fn get(&self) -> usize { self.size.load(Ordering::SeqCst) }
+    fn inc(&self) -> usize { self.size.fetch_add(1, Ordering::SeqCst) }
+    fn dec(&self) -> usize { self.size.fetch_sub(1, Ordering::SeqCst).saturating_sub(1) }
 }
 
-// Constructor
+// initializer
 pub async fn setup_queue() -> Queue {
-    let (queue, mut main_receiver, mut main_event_receive) = Queue::new();
+    let (queue, mut main_receiver, mut main_event_receiver) = Queue::new();
     let size_ref = queue.size.clone();
+    let queue_size = QueueSize { size: queue.size.clone() };
 
     let core_count = std::thread::available_parallelism()
         .map(|p| p.get())
@@ -179,7 +89,7 @@ pub async fn setup_queue() -> Queue {
 
     // TODO: tuned for dev env
     // needs tuning for a good middle ground of performance:UX
-    //let worker_count = core_count * 4;
+    // let worker_count = core_count * 4;
     let worker_count = 24;
 
     info!("Detected {core_count} CPU cores, spawning {worker_count} thumbnail workers");
@@ -187,34 +97,70 @@ pub async fn setup_queue() -> Queue {
     let mut worker_senders: Vec<mpsc::UnboundedSender<QueueItem>> =
     Vec::with_capacity(worker_count);
     for i in 0..worker_count {
+        let main_event_sender = queue.event_sender.clone();
         let (worker_sender, worker_receiver) = mpsc::unbounded_channel();
         worker_senders.push(worker_sender);
 
         let size_ref_clone = size_ref.clone();
         tokio::spawn(async move {
             debug!("Starting worker {i}");
-            QueueWorker::start(worker_receiver, size_ref_clone).await;
+            QueueWorker::start(worker_receiver, main_event_sender, size_ref_clone).await;
         });
     }
 
     tokio::spawn(async move {
         info!("Starting queue dispatcher.");
 
-        let mut next_worker_idx = 0;
-        while let Some(item) = main_event_receive.recv().await {
+        let mut processing: HashMap<String,QueueItem> = HashMap::new();
+        let mut blacklist: HashSet<String> = HashSet::new();
 
-            if let QueueMsg::ImageAdded = item {
-                if let Some(img) = main_receiver.recv().await {
-                    let worker_sender = &worker_senders[next_worker_idx];
-                    worker_sender.send(img)
-                        .unwrap_or_else(|e| error!("Failed to dispatch item to worker: {e}"));
-                    next_worker_idx = (next_worker_idx + 1) % worker_senders.len();
+        let mut next_worker_idx = 0;
+        while let Some(msg) = main_event_receiver.recv().await {
+
+            match msg {
+                QueueMsg::ImageAdded { file_hash: _ } => {
+                    if let Some(img) = main_receiver.recv().await {
+                        if blacklist.contains(&item_hash(&img)) {
+                            // TODO: message frontend
+                            continue;
+                        }
+
+                        if processing.contains_key(&item_hash(&img)) {
+                            debug!("Ignoring enqueue, already in queue: {}", &img.full_path.to_string_lossy());
+
+                            debug!("Decrementing queue from: {:?}", &size_ref);
+                            queue_size.dec();
+                            debug!("New queue size: {:?}", &size_ref);
+                            continue;
+                        }
+
+                        let worker_sender = &worker_senders[next_worker_idx];
+                        worker_sender.send(img)
+                            .unwrap_or_else(|e| error!("Failed to dispatch item to worker: {e}"));
+                        next_worker_idx = (next_worker_idx + 1) % worker_senders.len();
+                    }
+                }
+
+                QueueMsg::WorkerComplete { file_hash } => {
+                    // TODO: message frontend
+                    processing.remove(&file_hash);
+                    queue_size.dec();
+                }
+
+                QueueMsg::ImageFailed { file_hash } => {
+                    if !blacklist.contains(&file_hash) {
+                        blacklist.insert(file_hash);
+                    }
                 }
             }
-
         }
+
         info!("Queue dispatcher shutting down. Main receiver closed.");
     });
 
     queue
+}
+
+fn item_hash(item: &QueueItem) -> String {
+    format!("{}{}", &item.file_hash, &item.path_hash)
 }
